hacka_re_verifier/requirements.txt
# Core dependencies
jsbeautifier>=1.14.0
esprima>=4.0.1
beautifulsoup4>=4.10.0
playwright>=1.20.0
jinja2>=3.0.0
markdown>=3.3.0
pandas>=1.3.0
tabulate>=0.8.9
matplotlib>=3.4.0
pyyaml>=6.0

# Optional dependencies for PDF report generation
# weasyprint>=54.0

# Development dependencies
pytest>=7.0.0
pytest-playwright>=0.3.0
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.9.0
